agents/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
asyncio==3.4.3
